pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "google-api-python-client>=2.157.0",
    "google-auth>=2.37.0",
    "google-auth-oauthlib>=1.2.1",
]
